+++ credentials/model_provisioned_credential.go
@@ -20,7 +20,7 @@ var _ MappedNullable = &ProvisionedCredential{}
 
 // ProvisionedCredential struct for ProvisionedCredential
 type ProvisionedCredential struct {
-	Links *LinksHATEOAS `json:"_links,omitempty"`
+	Links *map[string]LinksHATEOASValue `json:"_links,omitempty"`
 	ClaimReference *ProvisionedCredentialClaimReference `json:"claimReference,omitempty"`
 	CreatedAt *time.Time `json:"createdAt,omitempty"`
 	Credential *ProvisionedCredentialCredential `json:"credential,omitempty"`
@@ -53,9 +53,9 @@ func NewProvisionedCredentialWithDefaults() *ProvisionedCredential {
 }
 
 // GetLinks returns the Links field value if set, zero value otherwise.
-func (o *ProvisionedCredential) GetLinks() LinksHATEOAS {
+func (o *ProvisionedCredential) GetLinks() map[string]LinksHATEOASValue {
 	if o == nil || IsNil(o.Links) {
-		var ret LinksHATEOAS
+		var ret map[string]LinksHATEOASValue
 		return ret
 	}
 	return *o.Links
@@ -63,7 +63,7 @@ func (o *ProvisionedCredential) GetLinks() LinksHATEOAS {
 
 // GetLinksOk returns a tuple with the Links field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *ProvisionedCredential) GetLinksOk() (*LinksHATEOAS, bool) {
+func (o *ProvisionedCredential) GetLinksOk() (*map[string]LinksHATEOASValue, bool) {
 	if o == nil || IsNil(o.Links) {
 		return nil, false
 	}
@@ -79,8 +79,8 @@ func (o *ProvisionedCredential) HasLinks() bool {
 	return false
 }
 
-// SetLinks gets a reference to the given LinksHATEOAS and assigns it to the Links field.
-func (o *ProvisionedCredential) SetLinks(v LinksHATEOAS) {
+// SetLinks gets a reference to the given map[string]LinksHATEOASValue and assigns it to the Links field.
+func (o *ProvisionedCredential) SetLinks(v map[string]LinksHATEOASValue) {
 	o.Links = &v
 }