+++ management/model_branding_settings.go
@@ -19,7 +19,7 @@ var _ MappedNullable = &BrandingSettings{}
 
 // BrandingSettings struct for BrandingSettings
 type BrandingSettings struct {
-	Links *LinksHATEOAS `json:"_links,omitempty"`
+	Links *map[string]LinksHATEOASValue `json:"_links,omitempty"`
 	// Specifies the resource’s unique identifier.
 	Id *string `json:"id,omitempty"`
 	Environment *ObjectEnvironment `json:"environment,omitempty"`
@@ -46,9 +46,9 @@ func NewBrandingSettingsWithDefaults() *BrandingSettings {
 }
 
 // GetLinks returns the Links field value if set, zero value otherwise.
-func (o *BrandingSettings) GetLinks() LinksHATEOAS {
+func (o *BrandingSettings) GetLinks() map[string]LinksHATEOASValue {
 	if o == nil || IsNil(o.Links) {
-		var ret LinksHATEOAS
+		var ret map[string]LinksHATEOASValue
 		return ret
 	}
 	return *o.Links
@@ -56,7 +56,7 @@ func (o *BrandingSettings) GetLinks() LinksHATEOAS {
 
 // GetLinksOk returns a tuple with the Links field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *BrandingSettings) GetLinksOk() (*LinksHATEOAS, bool) {
+func (o *BrandingSettings) GetLinksOk() (*map[string]LinksHATEOASValue, bool) {
 	if o == nil || IsNil(o.Links) {
 		return nil, false
 	}
@@ -72,8 +72,8 @@ func (o *BrandingSettings) HasLinks() bool {
 	return false
 }
 
-// SetLinks gets a reference to the given LinksHATEOAS and assigns it to the Links field.
-func (o *BrandingSettings) SetLinks(v LinksHATEOAS) {
+// SetLinks gets a reference to the given map[string]LinksHATEOASValue and assigns it to the Links field.
+func (o *BrandingSettings) SetLinks(v map[string]LinksHATEOASValue) {
 	o.Links = &v
 }