+++ management/model_identity_provider_apple.go
@@ -19,7 +19,7 @@ var _ MappedNullable = &IdentityProviderApple{}
 
 // IdentityProviderApple struct for IdentityProviderApple
 type IdentityProviderApple struct {
-	Links *LinksHATEOAS `json:"_links,omitempty"`
+	Links *map[string]LinksHATEOASValue `json:"_links,omitempty"`
 	// The description of the IdP.
 	Description *string `json:"description,omitempty"`
 	// The current enabled state of the IdP.
@@ -72,9 +72,9 @@ func NewIdentityProviderAppleWithDefaults() *IdentityProviderApple {
 }
 
 // GetLinks returns the Links field value if set, zero value otherwise.
-func (o *IdentityProviderApple) GetLinks() LinksHATEOAS {
+func (o *IdentityProviderApple) GetLinks() map[string]LinksHATEOASValue {
 	if o == nil || IsNil(o.Links) {
-		var ret LinksHATEOAS
+		var ret map[string]LinksHATEOASValue
 		return ret
 	}
 	return *o.Links
@@ -82,7 +82,7 @@ func (o *IdentityProviderApple) GetLinks() LinksHATEOAS {
 
 // GetLinksOk returns a tuple with the Links field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *IdentityProviderApple) GetLinksOk() (*LinksHATEOAS, bool) {
+func (o *IdentityProviderApple) GetLinksOk() (*map[string]LinksHATEOASValue, bool) {
 	if o == nil || IsNil(o.Links) {
 		return nil, false
 	}
@@ -98,8 +98,8 @@ func (o *IdentityProviderApple) HasLinks() bool {
 	return false
 }
 
-// SetLinks gets a reference to the given LinksHATEOAS and assigns it to the Links field.
-func (o *IdentityProviderApple) SetLinks(v LinksHATEOAS) {
+// SetLinks gets a reference to the given map[string]LinksHATEOASValue and assigns it to the Links field.
+func (o *IdentityProviderApple) SetLinks(v map[string]LinksHATEOASValue) {
 	o.Links = &v
 }