+++ management/model_identity_provider_paypal.go
@@ -19,7 +19,7 @@ var _ MappedNullable = &IdentityProviderPaypal{}
 
 // IdentityProviderPaypal struct for IdentityProviderPaypal
 type IdentityProviderPaypal struct {
-	Links *LinksHATEOAS `json:"_links,omitempty"`
+	Links *map[string]LinksHATEOASValue `json:"_links,omitempty"`
 	// The description of the IdP.
 	Description *string `json:"description,omitempty"`
 	// The current enabled state of the IdP.
@@ -69,9 +69,9 @@ func NewIdentityProviderPaypalWithDefaults() *IdentityProviderPaypal {
 }
 
 // GetLinks returns the Links field value if set, zero value otherwise.
-func (o *IdentityProviderPaypal) GetLinks() LinksHATEOAS {
+func (o *IdentityProviderPaypal) GetLinks() map[string]LinksHATEOASValue {
 	if o == nil || IsNil(o.Links) {
-		var ret LinksHATEOAS
+		var ret map[string]LinksHATEOASValue
 		return ret
 	}
 	return *o.Links
@@ -79,7 +79,7 @@ func (o *IdentityProviderPaypal) GetLinks() LinksHATEOAS {
 
 // GetLinksOk returns a tuple with the Links field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *IdentityProviderPaypal) GetLinksOk() (*LinksHATEOAS, bool) {
+func (o *IdentityProviderPaypal) GetLinksOk() (*map[string]LinksHATEOASValue, bool) {
 	if o == nil || IsNil(o.Links) {
 		return nil, false
 	}
@@ -95,8 +95,8 @@ func (o *IdentityProviderPaypal) HasLinks() bool {
 	return false
 }
 
-// SetLinks gets a reference to the given LinksHATEOAS and assigns it to the Links field.
-func (o *IdentityProviderPaypal) SetLinks(v LinksHATEOAS) {
+// SetLinks gets a reference to the given map[string]LinksHATEOASValue and assigns it to the Links field.
+func (o *IdentityProviderPaypal) SetLinks(v map[string]LinksHATEOASValue) {
 	o.Links = &v
 }