+++ management/model_sign_on_policy_action_id_first.go
@@ -19,7 +19,7 @@ var _ MappedNullable = &SignOnPolicyActionIDFirst{}
 
 // SignOnPolicyActionIDFirst struct for SignOnPolicyActionIDFirst
 type SignOnPolicyActionIDFirst struct {
-	Links *LinksHATEOAS `json:"_links,omitempty"`
+	Links *map[string]LinksHATEOASValue `json:"_links,omitempty"`
 	Condition *SignOnPolicyActionCommonConditionOrOrInner `json:"condition,omitempty"`
 	Environment *ObjectEnvironment `json:"environment,omitempty"`
 	// A string that specifies the sign-on policy assignment resource’s unique identifier.
@@ -58,9 +58,9 @@ func NewSignOnPolicyActionIDFirstWithDefaults() *SignOnPolicyActionIDFirst {
 }
 
 // GetLinks returns the Links field value if set, zero value otherwise.
-func (o *SignOnPolicyActionIDFirst) GetLinks() LinksHATEOAS {
+func (o *SignOnPolicyActionIDFirst) GetLinks() map[string]LinksHATEOASValue {
 	if o == nil || IsNil(o.Links) {
-		var ret LinksHATEOAS
+		var ret map[string]LinksHATEOASValue
 		return ret
 	}
 	return *o.Links
@@ -68,7 +68,7 @@ func (o *SignOnPolicyActionIDFirst) GetLinks() LinksHATEOAS {
 
 // GetLinksOk returns a tuple with the Links field value if set, nil otherwise
 // and a boolean to check if the value has been set.
-func (o *SignOnPolicyActionIDFirst) GetLinksOk() (*LinksHATEOAS, bool) {
+func (o *SignOnPolicyActionIDFirst) GetLinksOk() (*map[string]LinksHATEOASValue, bool) {
 	if o == nil || IsNil(o.Links) {
 		return nil, false
 	}
@@ -84,8 +84,8 @@ func (o *SignOnPolicyActionIDFirst) HasLinks() bool {
 	return false
 }
 
-// SetLinks gets a reference to the given LinksHATEOAS and assigns it to the Links field.
-func (o *SignOnPolicyActionIDFirst) SetLinks(v LinksHATEOAS) {
+// SetLinks gets a reference to the given map[string]LinksHATEOASValue and assigns it to the Links field.
+func (o *SignOnPolicyActionIDFirst) SetLinks(v map[string]LinksHATEOASValue) {
 	o.Links = &v
 }